+++ src/pages/login/components/WLogin.style.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
-import googleIcon from '../../../../img/login/google-icon.svg';
-import kakaoIcon from '../../../../img/login/kakao-icon.svg';
+import googleIcon from '../../../img/login/google-icon.svg';
+import kakaoIcon from '../../../img/login/kakao-icon.svg';
 
 export const Container = styled.div`
     box-sizing: border-box;
