+++ src/pages/login/components/WLogin.tsx
@@ -1,7 +1,7 @@
 import * as W from './WLogin.style';
-import LoginImage from '../../../../img/login/Login.png';
-import googleIcon from '../../../../img/login/google-icon.svg';
-import kakaoIcon from '../../../../img/login/kakao-icon.svg';
+import LoginImage from '../../../img/login/Login.png';
+import googleIcon from '../../../img/login/google-icon.svg';
+import kakaoIcon from '../../../img/login/kakao-icon.svg';
 
 interface LoginProps {
     google_login: () => void;
