+++ src/pages/login/components/WLoginComplete.style.tsx
@@ -1,28 +1,24 @@
 import styled from 'styled-components';
-import Check from '../../../../../img/login/Check.svg';
-import { Button } from '../../../../mypage/components/Common';
+import Check from '../../../img/login/Check.svg';
+import { Button } from '../../mypage/components/Common';
 
 export const Container = styled.div`
     display: flex;
-    flex-direction: column;
-    justify-content: flex-end;
-    align-items: center;
-    margin-bottom: 40px;
-
-    width: 100%;
     height: 100%;
+    justify-content: center;
+    align-items: center;
 
-    @media screen and (min-width: 768px) {
+    @media screen and (max-width: 767px) {
         display: none;
     }
 `;
 
 export const ItemBox = styled.div`
+    width: 290px;
+    margin: 0 auto;
     display: flex;
     flex-direction: column;
     align-items: center;
-    width: 100%;
-    margin-top: 10px;
     & > .img {
         width: 200px;
         height: 200px;
@@ -32,18 +28,15 @@ export const ItemBox = styled.div`
         margin-top: 24px;
         color: var(--Grey-900, #212224);
         text-align: center;
-
-        font-family: Pretendard;
         font-size: 24px;
-        font-style: normal;
         font-weight: 700;
-        line-height: 150%;
+        line-height: 150%; /* 36px */
     }
 `;
 
 export const LButton = styled(Button)`
-    margin-top: 237px;
-    width: 100%;
+    margin-top: 56px;
+    width: 182px;
     height: 56px;
     font-size: 20px;
 `;
