+++ src/pages/signUp/components/SchoolDropDown.tsx
@@ -2,7 +2,18 @@ import { useEffect, useState } from 'react';
 import Select, { ActionMeta, components } from 'react-select';
 import styled from 'styled-components';
 import { ReactComponent as Arrow } from '../../../img/arrow.svg';
+import useGetLocationUniv, {
+    IUniversity,
+} from '../../../query/get/useGetLocationUniv';
 import { schoolStyle } from './customSelectStyles';
+
+interface RegionCodeMap {
+    [key: string]: number;
+}
+
+interface UniversityNumber {
+    [key: string]: number;
+}
 
 export type RegionOptionType = {
     value: number;
@@ -27,44 +38,73 @@ const regionOptions: RegionOptionType[] = [
 
     { value: 25, label: '대전' },
     { value: 29, label: '세종' },
-    { value: 33, label: '충북/충남' },
+    { value: 33, label: '충북' },
+    { value: 34, label: '충남' },
     { value: 32, label: '강원' },
 
     { value: 21, label: '부산' },
-    { value: 26, label: '울산' },
     { value: 22, label: '대구' },
-    { value: 37, label: '경북/경남' },
+    { value: 37, label: '경남' },
+    { value: 38, label: '경북' },
 
-    { value: 24, label: '광주' },
-    { value: 35, label: '전북/전남' },
-    { value: 39, label: '제주' },
+    { value: 35, label: '전북' },
+    { value: 36, label: '전남' },
 ];
 
-// 대학 코드 번호 앞자리 2개가 지역 코드와 일치할 때만 필터링함
-const universityOptions: UniversityOptionType[] = [
-    { value: 1101, label: '서울대학교1' },
-    { value: 1102, label: '서울대학교2' },
-    { value: 1103, label: '서울대학교3' },
-    { value: 1104, label: '서울대학교4' },
-    { value: 1105, label: '서울대학교5' },
-    { value: 1106, label: '서울대학교6' },
-    { value: 1107, label: '서울대학교7' },
-    { value: 1108, label: '서울대학교8' },
-    { value: 1109, label: '서울대학교9' },
-    { value: 1110, label: '서울대학교10' },
-    { value: 1111, label: '서울대학교11' },
-    { value: 1112, label: '서울대학교12' },
-    { value: 3101, label: '경기대학교1' },
-    { value: 3102, label: '경기대학교2' },
-    { value: 2301, label: '인천대학교1' },
-    { value: 2302, label: '인천대학교2' },
-    { value: 2501, label: '대전대학교1' },
-    { value: 2502, label: '대전대학교2' },
-    { value: 2901, label: '세종대학교1' },
-    { value: 2902, label: '세종대학교2' },
-];
+const regionCodeMap: RegionCodeMap = {
+    서울: 11,
+    경기: 31,
+    인천: 23,
+    대전: 25,
+    세종: 29,
+    충북: 33,
+    충남: 34,
+    강원: 32,
+    부산: 21,
+    대구: 22,
+    경남: 37,
+    전북: 35,
+    전남: 36,
+    경북: 38,
+};
+
+const universityOptions: UniversityOptionType[] = [];
+export function findLabelByValue(value: number) {
+    const foundOption = universityOptions.find(
+        option => option.value === value,
+    );
+
+    if (!foundOption) {
+        return '해당하는 대학이 없습니다.';
+    }
+
+    return foundOption.label;
+}
 
 const SchoolDropDown = ({ onChange }: SchoolDropDownProps) => {
+    const universities: IUniversity[] = useGetLocationUniv({
+        activeTab: '전체',
+    });
+
+    const universityNumber: UniversityNumber = {};
+
+    // universityOptions 배열 생성
+    const universityOptions = universities.map(data => {
+        const regionCode = regionCodeMap[data.location];
+
+        // 해당 지역의 대학교 번호를 1 증가시킴 (존재하지 않는 경우 1로 초기화)
+        universityNumber[data.location] =
+            (universityNumber[data.location] || 0) + 1;
+
+        // 지역 코드와 대학교 번호를 결합하여 고유한 값을 생성
+        const value = regionCode * 100 + universityNumber[data.location];
+
+        return {
+            value,
+            label: data.universityName,
+        };
+    });
+
     // DropdownIndicator 컴포넌트 재정의(드롭다운 선택시 화살표 돌아가는)
     const DropdownIndicator = (props: any) => {
         return (
@@ -124,7 +164,7 @@ const SchoolDropDown = ({ onChange }: SchoolDropDownProps) => {
     };
 
     return (
-        <div>
+        <div style={{ marginBottom: '34px' }}>
             {filteredUniversity && (
                 <Select
                     options={filteredUniversity}
@@ -180,7 +220,7 @@ const CustomOption = styled.div`
     font-family: 'Pretendard';
     font-size: 16px;
     font-weight: 500;
-    width: 63px;
+    width: 110px;
     height: 32px;
     flex-shrink: 0;
     color: var(--grey-900, #212224);
@@ -200,4 +240,5 @@ const Border = styled.div`
     left: 142px;
     height: 347px;
     margin-top: 5px;
+    border-right: 1px solid var(--grey-400, #dcdfe3);
 `;
